# src/doc/swagger.yaml
openapi: 3.0.4
info:
  title: Blog API
  version: 2.0.0
  description: "API documentation for personal blogging platform"
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      summary: Get comments for a post
      description: |
        Retrieve a paginated list of comments for a post using optional cursor-based pagination.
        Supports sorting by latest, oldest, or top (most liked).  
        Returns a list of comments along with a cursor for the next page.
      tags: [Comments]
      parameters:
        - $ref: "#/components/parameters/PostIdParam"

        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 20
          required: false
          description: Number of comments to return.

        - $ref: "#/components/parameters/CursorParam"

        - $ref: "#/components/parameters/sortParam"

      responses:
        200:
          description: Paginated list of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  cursor:
                    type: string
                    nullable: true
                    description: Cursor to the next page (null if no more)
                    example: "64f1a8e9d3c9f3c6813cf28a"
                  comments:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          example: "64f1a8e9d3c9f3c6813cf28a"
                        body:
                          type: string
                          example: "This is a great post!"
                        likes:
                          type: integer
                          example: 5
                        replies: 
                          type: integer
                          example: 3
                        createdAt:
                          type: string
                          format: date-time
                        author:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: 672e4c0ee5a3844feb20db18
                            username:
                              type: string
                              example: "abdallah_dev"
                            profileImage:
                              type: string
                              example: "https://cdn.example.com/images/johndoe.png"
                              nullable: true
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples: 
                InvalidPostId:
                  summary: Invalid post Id
                  value: 
                    message: "Post id is not valid!"
                InvalidCursor:
                  summary: Invalid cursor
                  value: 
                    message: "Cursor does not exist"

        404:
          $ref: "#/components/responses/PostNotFound"

        500:
          $ref: "#/components/responses/ServerError"

    post:
      summary: Create a comment on a post
      operationId: createComment
      description: Allows an authenticated user to add a comment to a post.
      tags: [Comments]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  example: "This is a great post!"

      responses:
        201:
          $ref: "#/components/responses/Created"

        400:
          description: Invalid input (e.g. missing or invalid comment body)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    properties:
                      postId:
                        type: string
                        nullable: true
                        description: Error message about the post id parameter.
                        example: Invalid post id
                      body:
                        type: string
                        nullable: true
                        description: The comment's content
                        example: Comment must not be empty

        401:
          $ref: "#/components/responses/Unauthorized"

        404:
          $ref: "#/components/responses/PostNotFound"

        500:
          $ref: "#/components/responses/ServerError"
  /posts/{postId}/comments/{commentId}:
    patch:
      tags: [Comments]
      summary: Update a comment
      description: Updates the body of a specific comment. Only the comment's author is allowed to perform this action.
      operationId: updateComment
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  example: Updated comment content
      responses:
        200:
          $ref: "#/components/responses/Updated"
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    properties:
                      postId:
                        type: string
                        nullable: true
                        description: Error message about the post id parameter.
                        example: Invalid post id
                      commentId:
                        type: string
                        nullable: true
                        description: Error message about the comment id parameter.
                        example: Invalid comment id
                      body:
                        type: string
                        nullable: true
                        description: The comment's content
                        example: Comment must not be empty

        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: "#/components/responses/PostOrCommentNotFound"
        500:
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Delete a comment
      operationId: deleteComment
      description: |
        Deletes a comment if the user is the comment's author or has an admin/moderator role.
        Also deletes all associated replies and likes.

      tags: [Comments]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"

        - $ref: "#/components/parameters/CommentIdParam"

      responses:
        200:
          $ref: "#/components/responses/Deleted"

        400:
          description: Invalid postId or commentId
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid comment id"

        401:
          $ref: "#/components/responses/Unauthorized"

        403:
          $ref: "#/components/responses/Forbidden"

        404:
          $ref: "#/components/responses/PostOrCommentNotFound"

        500:
          $ref: "#/components/responses/ServerError"
  /posts/{postId}/comments/{commentId}/likes:
    post:
      tags: [Comments]
      summary: Like a comment
      description: Adds a like to a specific comment by the authenticated user. If already liked, does nothing.
      operationId: likeComment
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
      responses:
        200:
          $ref: "#/components/responses/Success"
        400:
          description: Invalid comment ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Comment id is not valid!
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: "#/components/responses/PostOrCommentNotFound"
        500:
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Unlike a comment
      operationId: unlikeComment
      description: |
        Removes the authenticated user's like from a specific comment.
        If the like doesn't exist, the operation is still considered successful (idempotent).

      tags: [Comments]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"

      responses:
        200:
          $ref: "#/components/responses/Success"

        400:
          description: Invalid post or comment ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment id is not valid!"

        401:
          $ref: "#/components/responses/Unauthorized"

        404:
          $ref: "#/components/responses/PostOrCommentNotFound"

        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Use the JWT token obtained from login or registration to access protected routes.
  parameters:
    PostIdParam:
      in: path
      name: postId
      required: true
      schema:
        type: string
        format: objectId
        example: "64f0341f3df05e12a3c3a401"
      description: ID of the post
    CommentIdParam:
      in: path
      name: commentId
      required: true
      schema:
        type: string
        format: objectId
        example: "64f0341f3df05212c3c3b402"
      description: ID of the comment
...
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
        nullable: true
        example: 64f0341f3df05e12a3c3a401
      description: Cursor to the next page (null if no more)
    sortParam:
      in: query
      name: sort
      schema:
        type: string
        enum: [latest, oldest, top]
        default: latest
      description: Sort order
  responses:
...
    Success:
      description: Success, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Success
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Created:
      description: Created, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Created
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Updated:
      description: Updated, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Updated
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Deleted:
      description: Deleted, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Deleted
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
...
    PostNotFound:
      description: Post not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message for post not found
                example: "Post not found"
    PostOrCommentNotFound:
      description: Post or comment not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Comment not found"
...
    Unauthorized:
      description: Unauthorized – user not authenticated or token is invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    Forbidden:
      description: Forbidden – user lacks permission (read description to know what roles are authorized)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message
...